backend/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==43.0.0
fastapi==0.112.2
greenlet==3.0.3
h11==0.14.0
idna==3.8
passlib==1.7.4
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==10.4.0
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dotenv==1.0.1
python-multipart==0.0.9
sniffio==1.3.1
SQLAlchemy==2.0.32
starlette==0.38.4
typing_extensions==4.12.2
uvicorn==0.30.6
